修复api调用报错的问题

+++ bakss-veeam/src/main/java/com/bakss/veeam/domain/login/LoginResponse.java
@@ -6,5 +6,5 @@ import lombok.Data;
 @Data
 public class LoginResponse {
     private String token;
-    private Integer expireAt;
+    private Long expireAt;
 }
